logread: Add support for output template
This patch let the users be able to define the output template, which
means one can decide how the output logs should look like by providing a
template. With this an user can easily integrate the output logs with
the existing log collecting services that only accept certain input
format.

It supports four pre-defined keywords: %message%, %priority%, %source%,
%timestamp%. These keywords will be replaced with the corresponding
information. Any other words in the template will be printed as-is.

Signed-off-by: Henry Chang <mr.changyuheng@gmail.com>